Ordering imports and moving code to match_utils
